+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - add abastract class of music elements and corresponding subcalsses(Note, Rest, Clef, TimeSigniture, KeySigniture)
 - add bpmController custom widget
 - add scroll for measures are playing
+- add scroll for notes are playing
 - show playing measure with a different color
 - show playing note with a different color